optimize include directives
