+++ STAY/code/common/asset/asset.cpp
@@ -1,5 +1,4 @@
 #include "asset.hpp"
-#include "utility/variantHelper.hpp"
 #include <filesystem>
 #include <cassert>
