+++ STAY/code/common/asset/type.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <filesystem>
-#include <variant>
 
 namespace stay 
 {
