+++ STAY/code/common/physics/rigidBody.cpp
@@ -1,6 +1,6 @@
 #include "rigidBody.hpp"
 #include "world.hpp"
-#include "../utility/convert.hpp"
+#include "utility/convert.hpp"
 
 namespace stay
 {
