+++ STAY/code/common/type/vector.hpp
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <type_traits>
-
 #include <glm/vec2.hpp>
 #include <glm/vec3.hpp>
