+++ STAY/code/common/utility/error.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
-#include <exception>
-#include <cassert>
+#include <stdexcept>
 
 namespace stay
 {
@@ -10,7 +9,7 @@ namespace stay
         template <typename Cond, typename Mes>
         void throwIfFalse(Cond&& cond, const Mes& message)
         {
-            if (static_cast<bool>(cond) == false)
+            if (!static_cast<bool>(cond))
                 throw std::runtime_error(message);
         }
     } // namespace utils
