+++ STAY/code/common/world/node.cpp
@@ -1,4 +1,3 @@
-#include <string>
 #include <cassert>
 
 #include "node.hpp"
