+++ STAY/code/game/component/player.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #include "../../common/physics/rigidBody.hpp"
-#include "../../common/ecs/manager.hpp"
 
 namespace stay
 {
@@ -23,19 +22,5 @@ namespace stay
         phys::RigidBody* hookBody;
 
         COMPONENT(player, moveStrength, jumpHeight, oppositeScale, airScale);
-    };
-
-    struct PlayerSystem 
-        : public ecs::System
-        , public ecs::StartSystem
-        , public ecs::UpdateSystem
-        , public ecs::InputSystem
-    {
-            PlayerSystem(ecs::Manager* manager);
-            void start() override;
-            void input(const sf::Event& event) override;
-            void update(float dt) override;
-        private:
-            bool mEntered;
     };
 } // namespace stay
