+++ STAY/code/game/system/list.hpp
@@ -1,10 +1,10 @@
 #pragma once
 
-#include "RawRenderSystem.hpp"
-#include "OrderedRenderSystem.hpp"
-#include "PhysicsDebugSystem.hpp"
-#include "PhysicsSystem.hpp"
-#include "hookSystem.hpp"
-#include "debugSystem.hpp"
-#include "../component/list.hpp"
-#include "dashSystem.hpp"
+#include "RawRenderSystem.hpp" // IWYU pragma: export
+#include "OrderedRenderSystem.hpp" // IWYU pragma: export
+#include "PhysicsDebugSystem.hpp" // IWYU pragma: export
+#include "PhysicsSystem.hpp" // IWYU pragma: export
+#include "playerSystem.hpp" // IWYU pragma: export
+#include "hookSystem.hpp" // IWYU pragma: export
+#include "debugSystem.hpp" // IWYU pragma: export
+#include "dashSystem.hpp" // IWYU pragma: export
