+++ STAY/code/game/system/playerSystem.hpp
@@ -1,0 +1,20 @@
+#pragma once
+
+#include "ecs/system.hpp"
+
+namespace stay 
+{
+    struct PlayerSystem 
+        : public ecs::System
+        , public ecs::StartSystem
+        , public ecs::UpdateSystem
+        , public ecs::InputSystem
+    {
+            PlayerSystem(ecs::Manager* manager);
+            void start() override;
+            void input(const sf::Event& event) override;
+            void update(float dt) override;
+        private:
+            bool mEntered;
+    };
+} // namespace stay
